// src/app/enum/pieces-names.enum.ts
export enum PiecesNamesEnum {
  WHITE_KING = 'whiteKing',
  WHITE_QUEEN = 'whiteQueen',
  WHITE_BISHOP = 'whiteBishop',
  WHITE_KNIGHT = 'whiteKnight',
  WHITE_ROOK = 'whiteRook',
  WHITE_PAWN = 'whitePawn',
  BLACK_KING = 'blackKing',
  BLACK_QUEEN = 'blackQueen',
  BLACK_BISHOP = 'blackBishop',
  BLACK_KNIGHT = 'blackKnight',
  BLACK_ROOK = 'blackRook',
  BLACK_PAWN = 'blackPawn'
}